Introduce OriginRequestProvider
Closes #35, closes #44

OriginRequestProvider is a Vert.x API type that unifies all the origin selection methods currently present as overloaded methods in HttpProxy.

Having a separate type allows:
- to create a better experience for developers using RxJava or Mutiny
- to provide a migration path (current methods are still supported)

Still, OriginRequestProvider being a FunctionalInterface, users can provide one in the form of a lambda.

OriginRequestProvider is a little different from the existing contract.
Instead of the proxied request as first parameter, it has the ProxyContext.

This allows to make a decision on the modifications interceptors can make (including when the request is an upgrade to WebSocket). See #44

It also allows interceptor and provider to exchange data via the ProxyContext attachments. See #35

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>